ubique funciones AsyncThunk en archivos de acciones correspondientes a cada slice en carpeta actions creada dentro de slices
