+++ src/domains/profesionales/components/FormularioPerfilProfesional.jsx
@@ -1,10 +1,9 @@
 import { Form } from "react-bootstrap";
-import imgUsuario from "../../../img/imgUsuario.png";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
-import { fetchWorkerForID } from "../../../slices/workersSlice";
+import { fetchWorkerForID } from "../../../slices/actions/workersActions";
 
 const FormularioPerfilProf = () => {
   const {
