+++ src/slices/registerSlice.js
@@ -1,5 +1,6 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import instance from "../api/api";
+import { createClient, createProfessional } from "./actions/registerActions";
 
 const initialState = {
     form: {
@@ -15,48 +16,6 @@ const initialState = {
     state: "idle",
     error: null
 };
-
-export const createProfessional = createAsyncThunk('register', async(_, { getState }) => {
-  try {
-    if (getState().register.form.type === 'Client') {
-      throw new Error('No se puede crear un cliente con este método');
-    }
-    
-    let sanitizedWorker = {
-      fullname: getState().register.form.fullname,
-      mail: getState().register.form.mail,
-      password: getState().register.form.password,
-      phone: getState().register.form.phone,
-      address: getState().register.form.adress,
-      img: getState().register.form.img ? getState().register.form.img : "",
-    }
-
-    console.log(sanitizedWorker)
-    
-    const response = await instance.post('/auth/jwt/register/Worker', sanitizedWorker);
-
-    return response.data.data;
-  } catch(error){
-    console.log(error);
-  }
-});
-
-export const createClient = createAsyncThunk('client', async(_, { getState }) => {
-  try {
-    const form = getState().register.form;
-
-    const { fullname, mail, password } = form
-
-    const response = await instance.post('/auth/jwt/register/', {
-      fullname,
-      mail,
-      password,
-    });
-    return response.data.data;
-  } catch(error){
-    console.log(error);
-  }
-});
 
 const registerSlice = createSlice({
     name: 'register',
